requirements.txt
numpy
pillow
imageio
numba
tqdm
scipy
matplotlib
streamlit
gradio
